+++ sdk/sdk/VKUploadPhotoRequest.m
@@ -23,7 +23,7 @@
 #import "VKUploadPhotoRequest.h"
 #import "VKApi.h"
 @implementation VKUploadPhotoRequest
-- (instancetype)initWithImage:(UIImage *)image parameters:(VKImageParameters *)parameters albumId:(int)albumId groupId:(int)groupId {
+- (instancetype)initWithImage:(UIImage *)image parameters:(VKImageParameters *)parameters albumId:(NSInteger)albumId groupId:(NSInteger)groupId {
 	self = [super init];
 	self.image            = image;
 	self.imageParameters  = parameters;
